Move to Python 3.14 and gate lint, format and types repo-wide
Raise the producer environment to 3.14: requires-python, the contracts
package, and every workflow interpreter. Adapter sources stay at 3.12
syntax because the pinned vLLM/SGLang images they overlay still run 3.12,
so the ruff target stays py312 and the shared three-d GLB validator keeps
matching the platform copy.

Add the repository ruff configuration and run the pinned 0.16.1 linter and
formatter over every authored file apart from two byte-exact vendored
trees, and add the platform's reviewed pyright baseline gate
(scripts/check-python-types) with a reason recorded for all 1580 accepted
errors.

Drop the digests that hashed a file shipped in the same commit, since a
source edit must not require hand-editing a digest the tooling owns: the
adapter patch-payload and Vision-Exp overlay tables, and three test
constants that digested an adapter tree. The upstream tokenizer encoding
and the downloaded-archive digests remain, and the packages and indexes
are regenerated for the reformatted sources.

+++ adapters/glm/mia-exl3-dflash2-dual/overlay/patch_glm_video_placeholders.py
@@ -32,7 +32,7 @@ def _align_timestamps(timestamps: list, t_groups: int) -> list:
     if n < t_groups:
         return timestamps + [timestamps[-1]] * (t_groups - n)
     last = n - 1
-    return [timestamps[int(round(i * last / (t_groups - 1)))] for i in range(t_groups)]
+    return [timestamps[round(i * last / (t_groups - 1))] for i in range(t_groups)]
 
 
 def _use_glm4v_timestamps(hf_processor) -> bool:
@@ -114,7 +114,7 @@ def _install_import_hook() -> None:
         applying = True
         try:
             apply()
-        except Exception as exc:
+        except Exception as exc:  # noqa: BLE001  (import hook must not break the host)
             print(f"glm53: video apply() failed: {exc!r}", file=sys.stderr)
         finally:
             applying = False
@@ -161,7 +161,7 @@ def _disable_gb10_persistent_topk() -> None:
 _install_import_hook()
 try:
     apply()
-except Exception:
+except Exception:  # noqa: BLE001, S110  (import-time best effort)
     pass
